feat: Make `int` a distinct primitive type, add expression classifier
This commit completes the foundational numeric type refactoring:

IR Type System:
- `int` is now primitiveType(name="int"), NOT primitiveType(name="number")
- Removed numericIntent from type level entirely
- IrPrimitiveTypeName includes "int" alongside "number", "string", "boolean"

Expression Converter:
- `as number` creates numericNarrowing with targetKind: "Double"
- Distinguishes explicit widening (42 as number) from plain literals (42)

Numeric Coercion Pass (TSN5110):
- Added classifyNumericExpr() to propagate numeric kinds through expressions
- Handles literals, identifiers, unary, binary, conditional, numericNarrowing
- Added hasExplicitDoubleIntent() for cast exemption detection
- Structural validation for arrays, ternaries, call arguments, returns

Emitter Updates:
- isIntegerType() checks primitiveType(name="int") instead of numericIntent
- hasInt32Proof() updated for int primitive

+++ packages/emitter/testcases/types/anonymous-objects/AnonymousObjects.cs
@@ -37,7 +37,7 @@ namespace TestCases.types.anonymousobjects
 
                 public static readonly object handler = new __Anon_AnonymousObjects_12_24 { id = 1, process = (double x) => x * 2 };
 
-                private static readonly double value = 100;
+                private static readonly int value = 100;
 
                 public static readonly object shorthand = new __Anon_AnonymousObjects_19_26 { value = value };
             }
